requirements.txt
openai
pypdf2
streamlit
langchain
langchain_community
python-dotenv

-e.